+++ src/main/java/com/lucendar/strm/common/StrmMsgs.java
@@ -15,14 +15,12 @@ public class StrmMsgs {
     public static final int STRM_MSG__OpenChannelReq = 4;
     public static final int STRM_MSG__KeepChannelReq = 5;
     public static final int STRM_MSG__ReleaseChannelsReq = 6;
-    public static final int STRM_MSG__StrmReportReq = 7;
-    public static final int STRM_MSG__QryServerStatReq = 8;
-    public static final int STRM_MSG__CheckChannelReq = 9;
-    public static final int STRM_MSG__GetTokensQry = 10;
-    public static final int STRM_MSG__ClosedChannelLogEntry = 11;
-    public static final int STRM_MSG__QryClosedLogReq = 12;
-    public static final int STRM_MSG__AvStoreLog = 13;
-    public static final int STRM_MSG__QryStrmActivityReq = 14;
-    public static final int STRM_MSG__QryStrmStatReq = 15;
-    public static final int STRM_MSG__QryStrmActSegReq = 16;
+    public static final int STRM_MSG__QryServerStatusReq = 7;
+    public static final int STRM_MSG__CheckChannelReq = 8;
+    public static final int STRM_MSG__ClosedChannelLogEntry = 9;
+    public static final int STRM_MSG__QryClosedLogReq = 10;
+    public static final int STRM_MSG__AvStoreLog = 11;
+    public static final int STRM_MSG__QryStrmActivityReq = 12;
+    public static final int STRM_MSG__QryStrmStatReq = 13;
+    public static final int STRM_MSG__QryStrmActSegReq = 14;
 }